Fixed MPI bug for taking curl of A
This is a hack; we need to communicate A into ghost zones across MPI
boundaries, but don't want to do all the bookkeeping work for this.  So we copy
A* arrays into the B*c arrays, and then call athena's bvals_mhd() function to
handle communication.  We then copy the B*c arrays back into A* and proceed to
take the curl.

*** Note that bvals_mhd() requires the function bvals_init(), which is called
AFTER problem() in main.c.  This is because boundary condition function
pointers are set in problem(), which bvals_init() depends on.  In order to call
bvals_mhd() from problem(), we had to move bvals_init() to before problem() in
main.c, which will break any user-specified boundary conditions set in
problem().  This is a dangerous hack in general, but hopefully OK for our
problem.

*** WARNING: switched order of bvals_init() and problem() in main.c, will break
for user specified boundary conditions ***