Fixed NAN problem for tangled field sims
For some reason when we fixed the curl across proccesors problem the4 magnetic field strength got a coupl orders of magnitude higher than previous sims.
I think this is what was giving us NANs
I honestly have no idea how or why this happened but I implemented an option through the problem block in athinput to set the order of magnitude of the field
strength.

Currently running a sim to test if this will work.
If I make no more commits in regards to this problem then it in fact worked. Fingers crossed!